refactor(utils): consolidate string primitives into utils/text/stringUtils
src/utils/core/stringCore.ts and src/utils/text/stringUtils.ts both held
generic string validation/formatting helpers with no documented rule for
which one a new helper belongs in — both were already imported side by
side from the same webview frontend files, so the split wasn't even
serving a browser-vs-Node boundary. Merge stringCore's exports into
stringUtils (the file AGENTS.md already documents as the home for text
utilities), keep the @utils/core barrel re-exporting the same symbols for
existing consumers, and repoint the handful of direct
`@utils/core/stringCore` imports. Also fixes AGENTS.md's stale
description of utils/core, which only mentioned debounce/delay despite
the directory holding type-guard, math, comparator, and URL helpers too.
