+++ packages/cli/src/runtime/apiStatus.ts
@@ -1,7 +1,7 @@
 import { platform } from '@platform/platform';
 import { toErrorMessage } from '@common/errors/errorMessage';
 import { API_PROVIDERS, lookupApiKeyOrigin } from '@model/apiProviders';
-import { formatPercent } from '@utils/core/stringCore';
+import { formatPercent } from '@utils/text/stringUtils';
 
 import {
   formatCliApiMode,
