+++ src/test-kernel/utils/text/StringUtils.vitest.ts
@@ -3,6 +3,8 @@ import { describe, expect, it } from 'vitest';
 
 // Local imports
 import {
+  formatCompactDuration,
+  formatCompactTokenCount,
   objectToLogString,
   pluralize,
   splitContentLines,
@@ -123,5 +125,38 @@ describe('Unicode-safe ellipsis helpers', () => {
     // Regression guard: slice(-(1-1)) === slice(0) once returned the whole tail.
     expect(tailWithEllipsis('abcdef', 1)).toBe('…');
     expect(tailWithEllipsis('abcdef', 0)).toBe('…');
+  });
+});
+
+describe('formatCompactDuration', () => {
+  it.each([
+    [-500, '0s'],
+    [0, '0s'],
+    [999, '0s'],
+    [1500, '1s'],
+    [59_000, '59s'],
+    [60_000, '1m'],
+    [185_000, '3m 5s'],
+    [3_600_000, '1h'],
+    [3_660_000, '1h 1m'],
+    [90_000_000, '1d 1h'],
+  ])('renders %i ms as %s', (ms, label) => {
+    expect(formatCompactDuration(ms)).toBe(label);
+  });
+});
+
+describe('formatCompactTokenCount', () => {
+  it.each([
+    [0, '0'],
+    [999, '999'],
+    [4096, '4096'],
+    [4097, '4k'],
+    [999_499, '999k'],
+    [999_500, '1.0M'],
+    [999_999, '1.0M'],
+    [1_000_000, '1.0M'],
+    [1_250_000, '1.3M'],
+  ])('renders %i tokens as %s', (tokens, label) => {
+    expect(formatCompactTokenCount(tokens)).toBe(label);
   });
 });
