+++ src/utils/core/index.ts
@@ -5,6 +5,10 @@
 // NOTE: pathCore is NOT re-exported here because it depends on Node.js 'path'.
 // This barrel is used by webview frontend code (browser context).
 // Backend code should import directly: import { ... } from '@utils/core/pathCore';
+//
+// String validation/formatting primitives live in @utils/text/stringUtils
+// (the single home for generic string helpers) and are re-exported here for
+// existing @utils/core consumers.
 export {
   isNonEmptyString,
   isString,
@@ -12,7 +16,7 @@ export {
   formatCompactTokenCount,
   formatDuration,
   serializeError,
-} from './stringCore';
+} from '../text/stringUtils';
 export {
   isObject,
   isFiniteNumber,
